backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic[email]==2.5.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
email-validator==2.1.0
matplotlib==3.8.2
plotly==5.18.0
pandas==2.1.4
kaleido==0.2.1
